Update submitting-issues.md
Where you say  "Take time to polish the format of your message so we'll enjoy reading it and working on it. " an alternative would be "...to enhance readability and make it enjoyable for us to work on it."

But I am not sure which version I prefer. I leave it up to you.